refactor: Move triggers.ts into TriggerSelection
